+++ tests/integration/proxy/v1/service/session/test_interaction.py
@@ -98,3 +98,28 @@ class InteractionTestCase(IntegrationTestCase):
                                      .interactions.list()
 
         self.assertIsNotNone(actual)
+
+    def test_delete_request(self):
+        self.holodeck.mock(Response(500, ''))
+
+        with self.assertRaises(TwilioException):
+            self.client.proxy.v1.services(sid="KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
+                                .sessions(sid="KCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
+                                .interactions(sid="KIaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").delete()
+
+        self.holodeck.assert_has_request(Request(
+            'delete',
+            'https://proxy.twilio.com/v1/Services/KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Sessions/KCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Interactions/KIaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+        ))
+
+    def test_delete_response(self):
+        self.holodeck.mock(Response(
+            204,
+            None,
+        ))
+
+        actual = self.client.proxy.v1.services(sid="KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
+                                     .sessions(sid="KCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
+                                     .interactions(sid="KIaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").delete()
+
+        self.assertTrue(actual)