+++ tests/integration/video/v1/test_room.py
@@ -46,6 +46,7 @@ class RoomTestCase(IntegrationTestCase):
                 "video_codecs": [
                     "VP8"
                 ],
+                "media_region": "us1",
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -92,6 +93,7 @@ class RoomTestCase(IntegrationTestCase):
                 "video_codecs": [
                     "VP8"
                 ],
+                "media_region": "us1",
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -164,6 +166,7 @@ class RoomTestCase(IntegrationTestCase):
                         "video_codecs": [
                             "VP8"
                         ],
+                        "media_region": "us1",
                         "url": "https://video.twilio.com/v1/Rooms/RM4070b618362c1682b2385b1f9982833c",
                         "links": {
                             "participants": "https://video.twilio.com/v1/Rooms/RM4070b618362c1682b2385b1f9982833c/Participants",
@@ -222,6 +225,7 @@ class RoomTestCase(IntegrationTestCase):
                 "video_codecs": [
                     "VP8"
                 ],
+                "media_region": "us1",
                 "end_time": "2015-07-30T20:00:00Z",
                 "duration": 10,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",