+++ twilio/rest/preview/hosted_numbers/authorization_document.py
@@ -35,13 +35,16 @@ class AuthorizationDocumentList(ListResource):
         self._solution = {}
         self._uri = '/AuthorizationDocuments'.format(**self._solution)
 
-    def stream(self, limit=None, page_size=None):
+    def stream(self, email=values.unset, status=values.unset, limit=None,
+               page_size=None):
         """
         Streams AuthorizationDocumentInstance records from the API as a generator stream.
         This operation lazily loads records as efficiently as possible until the limit
         is reached.
         The results are returned as a generator, so this operation is memory efficient.
 
+        :param unicode email: Email.
+        :param AuthorizationDocumentInstance.Status status: The Status of this AuthorizationDocument.
         :param int limit: Upper limit for the number of records to return. stream()
                           guarantees to never return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -54,16 +57,19 @@ class AuthorizationDocumentList(ListResource):
         """
         limits = self._version.read_limits(limit, page_size)
 
-        page = self.page(page_size=limits['page_size'])
+        page = self.page(email=email, status=status, page_size=limits['page_size'])
 
         return self._version.stream(page, limits['limit'], limits['page_limit'])
 
-    def list(self, limit=None, page_size=None):
+    def list(self, email=values.unset, status=values.unset, limit=None,
+             page_size=None):
         """
         Lists AuthorizationDocumentInstance records from the API as a list.
         Unlike stream(), this operation is eager and will load `limit` records into
         memory before returning.
 
+        :param unicode email: Email.
+        :param AuthorizationDocumentInstance.Status status: The Status of this AuthorizationDocument.
         :param int limit: Upper limit for the number of records to return. list() guarantees
                           never to return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -74,14 +80,16 @@ class AuthorizationDocumentList(ListResource):
         :returns: Generator that will yield up to limit results
         :rtype: list[twilio.rest.preview.hosted_numbers.authorization_document.AuthorizationDocumentInstance]
         """
-        return list(self.stream(limit=limit, page_size=page_size))
+        return list(self.stream(email=email, status=status, limit=limit, page_size=page_size))
 
-    def page(self, page_token=values.unset, page_number=values.unset,
-             page_size=values.unset):
+    def page(self, email=values.unset, status=values.unset, page_token=values.unset,
+             page_number=values.unset, page_size=values.unset):
         """
         Retrieve a single page of AuthorizationDocumentInstance records from the API.
         Request is executed immediately
 
+        :param unicode email: Email.
+        :param AuthorizationDocumentInstance.Status status: The Status of this AuthorizationDocument.
         :param str page_token: PageToken provided by the API
         :param int page_number: Page Number, this value is simply for client state
         :param int page_size: Number of records to return, defaults to 50
@@ -89,7 +97,13 @@ class AuthorizationDocumentList(ListResource):
         :returns: Page of AuthorizationDocumentInstance
         :rtype: twilio.rest.preview.hosted_numbers.authorization_document.AuthorizationDocumentPage
         """
-        params = values.of({'PageToken': page_token, 'Page': page_number, 'PageSize': page_size})
+        params = values.of({
+            'Email': email,
+            'Status': status,
+            'PageToken': page_token,
+            'Page': page_number,
+            'PageSize': page_size,
+        })
 
         response = self._version.page(
             'GET',